\name{BoutrosLab.plotting.general-package}
\Rdversion{1.1}
\alias{BoutrosLab.plotting.general-package}
\alias{BoutrosLab.plotting.general}
\docType{package}
\title{Boutros Lab general plotting functions}
\description{Functions for creating publication-quality plots}
\details{
\tabular{ll}{
Package: \tab BoutrosLab.plotting.general\cr
Type: \tab Package\cr
Version: \tab 5.3.4\cr
Date: \tab 2015-07-31\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
append.footnote();
colour.gradient();
covariates.grob();
create.colourkey();
create.barplot();
create.boxplot();
create.dendrogram();
create.densityplot();
create.dotmap();
create.heatmap();
create.hexbinplot();
create.histogram();
create.manhattanplot();
create.multiplot();
create.polygonplot();
create.qqplot.comparison();
create.qqplot.fit();
create.scatterplot();
create.segplot();
create.stripplot();
create.violinplot();
default.colours();
display.colours();
display.statistical.result();
force.colour.scheme();
general.at.final();
get.corr.key();
get.defaults();
get.line.breaks();
legend.grob();
scientific.notation();
show.available.palettes();
write.metadata();
write.plot();
gleason.score.colour();
}
\author{
Author: Paul C. Boutros <Paul.Boutros@oicr.on.ca>
Maintainer: Paul C. Boutros <Paul.Boutros@oicr.on.ca>
}
\keyword{ package }
